[BUILD FIXED] changed output to uint16_t

+++ include/init.h
@@ -13,7 +13,7 @@
 
 typedef struct opportunity
 {
-    uint8_t cv_in;
+    uint16_t cv_in;
     // NOTE: Currently CV only oscillates between
     // HIGH and LOW digital output values. It might be
     // better to also use voltage values for the output
@@ -23,5 +23,5 @@ typedef struct opportunity
 opportunity *OP_init(void);
 void *OP_free(void);
 
-void OP_read(opportunity *opt, uint8_t cv_in);
+void OP_read(opportunity *opt, uint16_t cv_in);
 void OP_process(opportunity *opt);
